# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Bundesdatenkrake",
    "version": "1.0.0",
    "description": "Über diese API können maschinenlesbare, versionierte Snapshots verschiedener Register (Quellen) bezogen werden."
  },
  "servers": [
    {
      "url": "https://www.krake.dev/api"
    },
    {
      "url": "http://localhost:3003/api"
    }
  ],
  "components": {
    "schemas": {
      "Resource": {
        "type": "object",
        "description": "Quelle",
        "properties": {
          "id": {
            "type": "string",
            "description": "Identifikator der Quelle",
            "example": "bafin-stimmrechte"
          },
          "name": {
            "type": "string",
            "description": "Name der Quelle",
            "example": "Stimmrechte BaFin"
          },
          "description": {
            "type": "string",
            "description": "Beschreibung der Quelle",
            "example": "Bedeutende Stimmrechtsanteile"
          },
          "sourceLink": {
            "type": "string",
            "description": "URL der Quelle",
            "example": "https://portal.mvp.bafin.de/database/AnteileInfo/"
          },
          "sourceName": {
            "type": "string",
            "description": "Für die Quelle verantwortliche Instanz",
            "example": "Bundesanstalt für Finanzdienstleistungsaufsicht"
          }
        }
      }
    }
  },
  "paths": {
    "/sources": {
      "get": {
        "summary": "Liste aller verfügbaren Quellen.",
        "responses": {
          "200": {
            "description": "Liste aller verfügbaren Quellen.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Resource"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/source": {
      "get": {
        "summary": "Verfügbare Dateien zu einer Quelle anzeigen",
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "description": "ID der Quelle",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Dateienliste zu einer Quelle",
            "content": {
              "application/json": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/file": {
      "get": {
        "summary": "Datei laden",
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "description": "ID der Quelle",
            "schema": {
              "type": "string"
            },
            "required": true
          },
          {
            "in": "query",
            "name": "file",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dateiinhalt als JSON"
          }
        }
      }
    },
    "/schema": {
      "get": {
        "summary": "Schema einer Datei laden",
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "description": "ID der Quelle",
            "schema": {
              "type": "string"
            },
            "required": true
          },
          {
            "in": "query",
            "name": "file",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Schema als JSON-Schema"
          }
        }
      }
    }
  }
}